Refactor project structure
refactored the sturcture of code files
