+++ src/main/java/com/emma/app/action/AttendanceAction.java
@@ -1,25 +1,15 @@
-package com.emma.action;
+package com.emma.app.action;
 
 import com.emma.app.bean.AttendanceBean;
 import com.emma.app.bean.AttendanceBeanI;
-import com.emma.app.bean.EmployeeBean;
-import com.emma.app.bean.EmployeeBeanI;
-import com.emma.app.model.entity.Attendance;
-import com.emma.app.model.entity.Employee;
-import com.emma.app.model.entity.EmployeeRole;
-import com.emma.database.Database;
+import com.emma.app.model.Attendance;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
-import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 import javax.servlet.http.HttpSession;
 import java.io.IOException;
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Formatter;
 
 @WebServlet("/add-attendance")
 public class AttendanceAction extends BaseAction {
