+++ src/main/java/com/emma/app/action/EmployeeAction.java
@@ -1,9 +1,9 @@
-package com.emma.action;
+package com.emma.app.action;
 
 import com.emma.app.bean.EmployeeBean;
 import com.emma.app.bean.EmployeeBeanI;
-import com.emma.app.model.entity.Employee;
-import com.emma.app.view.html.HtmlComponent;
+import com.emma.app.model.Employee;
+import com.emma.app.view.helper.HtmlComponent;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
