+++ src/main/java/com/emma/app/bean/AttendanceBeanI.java
@@ -1,6 +1,6 @@
 package com.emma.app.bean;
 
-import com.emma.app.model.entity.Attendance;
+import com.emma.app.model.Attendance;
 
 import javax.servlet.http.HttpServletRequest;
