+++ src/main/java/com/emma/app/bean/EmployeeBean.java
@@ -1,8 +1,7 @@
 package com.emma.app.bean;
 
-import com.emma.app.model.entity.Attendance;
-import com.emma.app.model.entity.Employee;
-import com.emma.app.view.html.HtmlComponent;
+import com.emma.app.model.Employee;
+import com.emma.app.view.helper.HtmlComponent;
 import com.emma.database.Database;
 
 import java.io.Serializable;
