+++ src/main/java/com/emma/app/bean/UserBean.java
@@ -1,6 +1,6 @@
 package com.emma.app.bean;
 
-import com.emma.app.model.entity.User;
+import com.emma.app.model.User;
 import com.emma.database.Database;
 
 import java.io.Serializable;
