+++ src/main/java/com/emma/app/bean/UserBeanI.java
@@ -1,6 +1,6 @@
 package com.emma.app.bean;
 
-import com.emma.app.model.entity.User;
+import com.emma.app.model.User;
 
 public interface UserBeanI {
     boolean register(User user);
