+++ src/main/java/com/emma/app/model/EmployeeRole.java
@@ -1,4 +1,4 @@
-package com.emma.app.model.entity;
+package com.emma.app.model;
 
 public enum EmployeeRole {
     DEVELOPER, QA, DEVOPS;
