+++ src/main/java/com/emma/app/view/helper/Menu.java
@@ -1,4 +1,4 @@
-package com.emma.app.view.toolbar;
+package com.emma.app.view.helper;
 
 public interface Menu {
     String menu(int activeMenu);
