+++ src/main/java/com/emma/app/view/helper/MenuLinkStatus.java
@@ -1,4 +1,4 @@
-package com.emma.app.model.view;
+package com.emma.app.view.helper;
 
 public enum MenuLinkStatus {
     ACTIVE, NOT_ACTIVE
