+++ src/main/java/com/emma/app/view/helper/MyTableColHeader.java
@@ -1,4 +1,4 @@
-package com.emma.app.view.html;
+package com.emma.app.view.helper;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
