+++ src/main/webapp/app/index.jsp
@@ -1,5 +1,5 @@
 <%@ page isELIgnored="false" %>
-<jsp:useBean id="topBar" class="com.emma.app.view.toolbar.TopBar" scope="request"/>
+<jsp:useBean id="topBar" class="com.emma.app.view.helper.TopBar" scope="request"/>
 <jsp:useBean id="menuBean" class="com.emma.app.displaybean.MenuBean" scope="request"/>
